% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_simulation_type1.R
\name{do_simulation_type1}
\alias{do_simulation_type1}
\title{Do Type 1 Simulation}
\usage{
do_simulation_type1(df)
}
\arguments{
\item{df}{data frame containing columns named by parameter}
}
\value{
data frame
}
\description{
In this simulation data is simulated from scratch for each iteration, ie no data is shared
between simulations
}
\examples{
NULL
}
