% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_simulation_type2.R
\name{do_simulation_type2}
\alias{do_simulation_type2}
\title{Do Type 2 Simulation}
\usage{
do_simulation_type2(df)
}
\arguments{
\item{df}{data frame containing columns named by parameter}
}
\value{
data frame
}
\description{
In this simulation data is simulated from scratch for each iteration, ie no data is shared
between simulations.  Test level statistics are calculated.
}
\examples{
NULL
}
