% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_method.R
\name{lm_method}
\alias{lm_method}
\title{LM method}
\usage{
lm_method(df, pIC50_col = "pIC50")
}
\arguments{
\item{df}{data frame including gene and pIC50 columns}

\item{pIC50_col}{name of the pIC50 column (default is pIC50)}
}
\value{
data frame of results
}
\description{
Simple linear regression and return output in standardised form
}
\examples{
NULL
}
