% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit.R
\name{nlme_fit}
\alias{nlme_fit}
\title{Mixed Effects Model Fit}
\usage{
nlme_fit(df)
}
\arguments{
\item{df}{A data frame with column names resp, conc and cell_id.}
}
\value{
an object of class `nlme` representing the nonlinea mixed-effects model fit.
}
\description{
Fit a dose response curve using \pkg{nlme}.  Estimated pIC50 values can be extracted from
the fit using the \code{\link{nlme_extract}} function.
}
\examples{
data(drc_data_small)
fit <- nlme_fit(drc_data_small)
fit
nlme_extract(fit)
}
