% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_gene_fit.R
\name{nlme_gene_fit}
\alias{nlme_gene_fit}
\title{Mixed Effects Gene Model Fit}
\usage{
nlme_gene_fit(df)
}
\arguments{
\item{df}{A data frame with column names resp, conc, gene and cell_id.}
}
\value{
an object of class `nlme` representing the nonlinear mixed-effects model fit.
}
\description{
Fit a dose response curve using \pkg{nlme} and including the genetic effect as a covariate
}
\examples{
data(drc_data_small)
fit <- nlme_gene_fit(drc_data_small)
fit
nlme_gene_extract(fit)
}
