% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_gene_method.R
\name{nlme_gene_method}
\alias{nlme_gene_method}
\title{NLME gene method}
\usage{
nlme_gene_method(df)
}
\arguments{
\item{df}{nested data frame with cell_id, gene, pIC50 and dose response data}
}
\value{
data frame of results
}
\description{
Estimate the genetic parameter directly within the non-linear model
}
\examples{
NULL
}
