% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_lm_method.R
\name{nlme_lm_method}
\alias{nlme_lm_method}
\title{NLME + LM method}
\usage{
nlme_lm_method(df)
}
\arguments{
\item{df}{nested data frame with cell_id, gene, pIC50 and dose response data}
}
\value{
data frame of results
}
\description{
Estimate pIC50 values for each cell line using non-linear mixed model
then do linear regression
}
\examples{
NULL
}
