% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_fit.R
\name{nls_fit}
\alias{nls_fit}
\title{Non-linear Least Squares Fit}
\usage{
nls_fit(df)
}
\arguments{
\item{df}{A data frame with column names resp, conc and cell_id.}
}
\value{
an object of class `nls` representing the nonlinear least squares model fit.
}
\description{
Fit a dose response curve using the Levenberg-Marquardt Nonlinear Least-Squares squares
algorithm from the \pkg{minpack.lm} package.  Estimated pIC50 values can be extracted from
the fit using the \code{\link{nls_extract}} function.
}
\examples{
data(drc_data_small)
fit <- nls_fit(drc_data_small)
fit
broom::tidy(fit)
}
