% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_lm_method.R
\name{nls_lm_method}
\alias{nls_lm_method}
\title{NLS + LM method}
\usage{
nls_lm_method(df)
}
\arguments{
\item{df}{nested data frame with cell_id, gene, pIC50 and dose response data}
}
\value{
data frame of results
}
\description{
Estimate pIC50 for each cell line using non-linear least squares then do linear
regression
}
\examples{
data("sim_ex1_data")
dat <- sim_ex1_data$data[[1]]
dat
nls_lm_method(dat)

library(tidyverse)
purrr::map(sim_ex1_data$data, nls_lm_method) \%>\% dplyr::bind_rows()
}
