% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgxsim_example_data.R
\name{pgxsim-data}
\alias{pgxsim-data}
\alias{cell_lines_small}
\alias{cell_lines_large}
\alias{pIC50_data_small}
\alias{pIC50_data_large}
\alias{drc_data_small}
\alias{drc_data_large}
\alias{parallel_res_df}
\alias{sim_ex1_df}
\alias{sim_ex1_data}
\title{Example datasets for the pgxsim package}
\description{
The pgxsim package has the following example datasets:
\describe{
 \item{cell_lines_small}{10 cell lines simulated with a 20\% frequency mutation}
 \item{cell_lines_large}{100 cell lines simulated with a 20\% frequency mutation}
 \item{pIC50_data_small}{pIC50 data simulated for \link{cell_lines_small} (10 cell lines) with a genetic coefficient of 1}
 \item{pIC50_data_large}{pIC50 data simulated for \link{cell_lines_large} (100 cell lines) with a genetic coefficient of 1}
 \item{drc_data_small}{Dose response data simulated for \link{pIC50_data_small} (10 cell lines)}
 \item{drc_data_large}{Dose response data simulated for \link{pIC50_data_large} (100 cell lines)}
 \item{parallel_res_df}{Parallelised simulations results dataset for vignette}
 \item{sim_ex1_df}{Data frame defining example set of simulations}
 \item{sim_ex1_data}{Simulated dose response data from sim_ex1_df}
 }
}
