% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgxsim_documentation.R
\docType{package}
\name{pgxsim-package}
\alias{pgxsim-package}
\alias{pgxsim}
\title{Simulate Pharmacogenomics Data}
\description{
Package to simulate Pharmacogenomics dose response data
}
\details{
To open the vignette type \code{vignette("Overview_of_pgxsim")}
}
\author{
Phil Chapman <phil.chapman@cruk.manchester.ac.uk>
}
