% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cell_lines.R
\name{sim_cell_lines}
\alias{sim_cell_lines}
\title{Simulate Cell Lines}
\usage{
sim_cell_lines(n, type = "d", prop = 0.2, prho = 0.9, mu = 0,
  sd = 0.1, beta = 1, lb = -4, ub = Inf, sd_prop = 0, sd_add = 0)
}
\arguments{
\item{n}{Number of cell lines to simulate}

\item{type}{Whether genetic covariate is a discrete (d) or continuous (c) variable}

\item{prop}{Proportion of cell lines with a discrete characteristic}

\item{prho}{value of correlation coefficient (0-1 for continuous, ignored for discrete)}

\item{mu}{mean of the truncated normal distribution}

\item{sd}{standard deviation of the truncated normal distribution}

\item{beta}{value of the genetic coefficient to recover - note that this is on a log scale}

\item{lb}{lower bound of the pIC50 distribution}

\item{ub}{upper bound of the pIC50 distribution}

\item{sd_prop}{Cell line specific proportional dose response variance}

\item{sd_add}{Cell line specific additive dose response variance}
}
\value{
A tibble containing cell line data
}
\description{
Simulate cell lines with predefined genetic and drug response characteristics.
}
\examples{
library(dplyr)
set.seed(10000)
sim_cell_lines(n=10, type='d', prop=0.2, mu=1, sd=0.1, beta=1)
sim_cell_lines(n=10, type='c', mu=1, sd=0.1, beta=1, prho=0.6)
}
