% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dose_response.R
\name{sim_dose_response}
\alias{sim_dose_response}
\title{Simulate Dose Response Data}
\usage{
sim_dose_response(pIC50 = 1, lb = 0.001, ub = 30, ndoses = 10,
  nreps = 3, sd_prop = 0.1, sd_add = 0.1)
}
\arguments{
\item{pIC50}{The pIC50 for the cell line}

\item{lb}{Lowest concentration the cell line is tested at}

\item{ub}{Highest concentration the cell line is tested at}

\item{ndoses}{Number of doses}

\item{nreps}{Numer of replicates}

\item{sd_prop}{Proportional standard deviation}

\item{sd_add}{Additive standard deviation}
}
\value{
A data frame of dose response values
}
\description{
Simulates dose response data for a cell line given a pIC50, upper and lower bounds, number
of doses, number of replicates per dose and additive and proportional variance
}
\examples{
library(ggplot2)
#no error
ex1 <- sim_dose_response(1, 0.001, 30, 10, 3, 0, 0)
ex1
qplot(conc, resp, data=ex1) + scale_x_log10()

#some proportional error
ex2 <- sim_dose_response(1, 0.001, 30, 10, 3, 0.1, 0)
ex2
qplot(conc, resp, data=ex2) + scale_x_log10()

#some additive error
ex3 <- sim_dose_response(1, 0.001, 30, 10, 3, 0, 0.1)
ex3
qplot(conc, resp, data=ex3) + scale_x_log10()

#both error types
ex4 <- sim_dose_response(1, 0.001, 30, 10, 3, 0, 0.1)
ex4
qplot(conc, resp, data=ex3) + scale_x_log10()
}
