% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_apply.R
\name{subset_apply}
\alias{subset_apply}
\title{Subset and Apply Function}
\usage{
subset_apply(k, df, my_fun)
}
\arguments{
\item{k}{batch id to select}

\item{df}{data frame to subset}

\item{my_fun}{function to apply}
}
\description{
Subset a data frame and apply a function to that portion of it
}
\examples{
NULL
}
