% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_simulation_type2.R
\name{do_simulation_type2}
\alias{do_simulation_type2}
\title{Do Type 2 Simulation}
\usage{
do_simulation_type2(df, output = "results")
}
\arguments{
\item{df}{data frame containing columns named by parameter}

\item{output}{specify what to return: `results` for a data frame of results (default),
`data` to simulate data only or 'both' for a list containing both results and data}
}
\value{
data frame
}
\description{
In this simulation data is simulated from scratch for each iteration, ie no data is shared
between simulations.  Test level statistics are calculated.
}
\examples{

\dontrun{
data(sim_ex1_df)
do_simulation_type2(sim_ex1_df) #just return results
do_simulation_type2(sim_ex1_df, output='data') #just simulate dose response data
do_simulation_type2(sim_ex1_df, output='both') #return results and simulated data
}
}
