% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_extract.R
\name{nlme_extract}
\alias{nlme_extract}
\title{Extract pIC50s From nlme}
\usage{
nlme_extract(fit)
}
\arguments{
\item{fit}{}
}
\value{
data frame of values
}
\description{
Extract the pIC50 values from an nlme fit.
}
\examples{
#
}
