% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_extract.R
\name{nls_extract}
\alias{nls_extract}
\title{Extract pIC50s From nls}
\usage{
nls_extract(fit)
}
\arguments{
\item{fit}{output from \code{\link{nls_fit}}}
}
\value{
data frame of values
}
\description{
Extract the pIC50 values from an nls fit.
}
\examples{
library(tidyverse)
data(drc_data_small)
cl_data_small <- dplyr::group_by(drc_data_small, cell_id) \%>\% tidyr::nest()
fit <- nls_fit(cl_data_small$data[[1]])
fit
nls_extract(fit)
}
