% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_fit.R
\name{nls_fit}
\alias{nls_fit}
\title{Non-linear Least Squares Fit}
\usage{
nls_fit(df)
}
\arguments{
\item{df}{A data frame with column names resp, conc and cell_id.}
}
\value{
an object of class `nls` representing the nonlinear least squares model fit.
}
\description{
Fit a dose response curve using the Levenberg-Marquardt Nonlinear Least-Squares squares
algorithm from the \pkg{minpack.lm} package.  Estimated pIC50 values can be extracted from
the fit using the \code{\link{nls_extract}} function.
}
\details{
NOTE: sometimes the fit will fail, in which case null is returned
}
\examples{
data(drc_data_small)
cl1_data <- dplyr::filter(drc_data_small, cell_id==1)
fit <- nls_fit(cl1_data)
fit
broom::tidy(fit)  #on log-normal scale
nls_extract(fit)  #convenience function converts scales automatically
}
