% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cell_lines.R
\name{sim_cell_lines}
\alias{sim_cell_lines}
\title{Simulate Cell Lines}
\usage{
sim_cell_lines(n, type = "d", prop = 0.2, sd_prop = 0, sd_add = 0)
}
\arguments{
\item{n}{Number of cell lines to simulate}

\item{type}{Whether genetic covariate is a discrete (d) or continuous (c) variable}

\item{prop}{Proportion of cell lines with a discrete characteristic}

\item{sd_prop}{Cell line specific proportional dose response variance}

\item{sd_add}{Cell line specific additive dose response variance}
}
\value{
A tibble containing cell line data
}
\description{
Simulate a cell lines with predefined genetic characteristics
}
\examples{
library(dplyr)
set.seed(10000)
sim_cell_lines(n=10, type='d', prop=0.2)
sim_cell_lines(n=10, type='c')
}
